+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -290,6 +290,7 @@
   _(BOOLEAN                     , DisableRtcChecksum          ,     , FALSE  , ()) \
   _(BOOLEAN                     , ExtendBTFeatureFlags        ,     , FALSE  , ()) \
   _(BOOLEAN                     , ExternalDiskIcons           ,     , FALSE  , ()) \
+  _(BOOLEAN                     , ForceAquantiaEthernet       ,     , FALSE  , ()) \
   _(BOOLEAN                     , ForceSecureBootScheme       ,     , FALSE  , ()) \
   _(BOOLEAN                     , IncreasePciBarSize          ,     , FALSE  , ()) \
   _(BOOLEAN                     , LapicKernelPanic            ,     , FALSE  , ()) \