+++ Changelog.md
@@ -3,6 +3,7 @@ OpenCore Changelog
 #### v0.8.0
 - Added support for early log preservation
 - Switched to Python 3 in scripts (use `python /path/to/script` to force Python 2)
+- Added `ForceAquantiaEthernet` for Aquantia AQtion AQC-107s based 10GbE network cards support, thx @Mieze
 
 #### v0.7.9
 - Added auto-detect `macOS Installer` volume name for use when `.disk_label` file cannot be displayed